refactor: Update language button component import and CV links in en.json and es.json
