+++ src/components/index.ts
@@ -19,7 +19,7 @@ export * from './ui/bento-grid/BentoGrid';
 export * from './ui/button/Button';
 export * from './ui/formatted-date/FormattedDate';
 export * from './ui/highlight-text/HighlightText';
-export * from './ui/laguange-button/LaguageButton';
+export * from './ui/language-button/LanguageButton';
 export * from './ui/spotlight-effect/Spotlight';
 export * from './ui/text-generate-effect/TextGenerateEffect';
 export * from './ui/timeline/Timeline';
