+++ src/components/navbar/Navbar.tsx
@@ -14,7 +14,7 @@ import { useUIStore } from '@/store';
 import { languages } from '@/locales';
 import { titleFont } from '@/config';
 import { navItemsProps } from '@/interfaces';
-import { LanguageButton } from '../ui/laguange-button/LaguageButton';
+import { LanguageButton } from '../ui/language-button/LanguageButton';
 
 export const Navbar = ({ navItems, href }: navItemsProps) => {
   const openMenu = useUIStore((state) => state.openSideMenu);
