+++ src/locales/en.json
@@ -22,7 +22,7 @@
     }
   ],
 
-  "cvLink": "https://res.cloudinary.com/dmlpgks2h/image/upload/fl_attachment/v1719200911/Portfolio/cv-en-marco-rodriguez.pdf",
+  "cvLink": "https://res.cloudinary.com/dmlpgks2h/image/upload/fl_attachment/v1724264175/Portfolio/cv-en-marco-rodriguez.pdf",
 
   "headerInfo": {
     "words": [
