+++ src/locales/es.json
@@ -22,7 +22,7 @@
     }
   ],
 
-  "cvLink": "https://res.cloudinary.com/dmlpgks2h/image/upload/fl_attachment/v1720633664/Portfolio/cv-es-marco-rodriguez.pdf",
+  "cvLink": "https://res.cloudinary.com/dmlpgks2h/image/upload/fl_attachment/v1724263943/Portfolio/cv-es-marco-rodriguez.pdf",
 
   "headerInfo": {
     "words": [
